Change datatype for total_amount in Order entity

+++ src/main/java/com/gr1tEnt/models/Order.java
@@ -2,6 +2,7 @@ package com.gr1tEnt.models;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
@@ -14,6 +15,6 @@ public class Order {
     private UUID customer_id;
     private LocalDate order_date;
     private OrderStatus status;
-    private double total_amount;
+    private BigDecimal total_amount;
     private String shipping_address;
 }
